+++ packages/editor/src/types.ts
@@ -13,16 +13,130 @@ import type {
 /**
  * WordPress dependencies
  */
-import type { BlockInstance } from '@wordpress/blocks';
+import type { BlockInstance, BlockEditProps } from '@wordpress/blocks';
 
-export type BulkOptimizationAttachmentData = Pick<
-	Attachment,
-	'id' | 'url' | 'mexp_filesize' | 'mexp_filename'
-> & {
-	posterUrl: Attachment[ 'url' ];
-	clientId: BlockInstance[ 'clientId' ];
-	isOptimized: boolean;
+export type ImageBlock = BlockInstance< {
+	id: number;
+	url: string;
+	caption: string;
+	alt: string;
+} > & { name: 'core/image' };
+
+export type AudioBlock = BlockInstance< {
+	id: number;
+	src: string;
+} > & { name: 'core/audio' };
+
+export type VideoBlock = BlockInstance< {
+	id: number;
+	src: string;
+	poster: string;
+	muted: boolean;
+	caption: string;
+	tracks: Array< {
+		src?: string;
+		label?: string;
+		srcLang?: string;
+		kind: 'subtitles' | 'captions';
+	} >;
+	controls: boolean;
+	loop: boolean;
+	autoplay: boolean;
+	playsInline: boolean;
+} > & { name: 'core/video' };
+
+export type MediaTextBlock = BlockInstance< {
+	mediaId: number;
+	mediaUrl: string;
+	mediaType: string;
+} > & { name: 'core/media-text' };
+
+export type GalleryBlock = BlockInstance< {
+	images: Array< {
+		id: number;
+		url: string;
+		alt: string;
+		caption: string;
+	} >;
+} > & { name: 'core/gallery' };
+
+export type CoverBlock = BlockInstance< {
+	id: number;
+	url: string;
+	useFeaturedImage: boolean;
+	backgroundType: string;
+} > & { name: 'core/cover' };
+
+export type PostFeaturedImageBlock = BlockInstance< {} > & {
+	name: 'core/post-featured-image';
+};
+
+export type SiteLogoBlock = BlockInstance< {
+	shouldSyncIcon: boolean;
+} > & {
+	name: 'core/site-logo';
+};
+export type EmbedBlock = BlockInstance< {
+	url: string;
+	providerNameSlug: string;
+} > & {
+	name: 'core/embed';
+};
+
+export type MediaPanelProps =
+	| ( ImageBlock &
+			Pick<
+				BlockEditProps< ImageBlock[ 'attributes' ] >,
+				'setAttributes' | 'className'
+			> )
+	| ( VideoBlock &
+			Pick<
+				BlockEditProps< VideoBlock[ 'attributes' ] >,
+				'setAttributes' | 'className'
+			> )
+	| ( AudioBlock &
+			Pick<
+				BlockEditProps< AudioBlock[ 'attributes' ] >,
+				'setAttributes' | 'className'
+			> )
+	| ( GalleryBlock &
+			Pick<
+				BlockEditProps< GalleryBlock[ 'attributes' ] >,
+				'setAttributes' | 'className'
+			> )
+	| ( MediaTextBlock &
+			Pick<
+				BlockEditProps< MediaTextBlock[ 'attributes' ] >,
+				'setAttributes' | 'className'
+			> )
+	| ( CoverBlock &
+			Pick<
+				BlockEditProps< CoverBlock[ 'attributes' ] >,
+				'setAttributes' | 'className'
+			> )
+	| ( PostFeaturedImageBlock &
+			Pick<
+				BlockEditProps< PostFeaturedImageBlock[ 'attributes' ] >,
+				'setAttributes' | 'className'
+			> )
+	| ( SiteLogoBlock &
+			Pick<
+				BlockEditProps< SiteLogoBlock[ 'attributes' ] >,
+				'setAttributes' | 'className'
+			> )
+	| ( EmbedBlock &
+			Pick<
+				BlockEditProps< EmbedBlock[ 'attributes' ] >,
+				'setAttributes' | 'className'
+			> );
+
+export type BulkOptimizationAttachmentData = {
+	id: number;
+	url: string;
+	filesize: number | null;
+	filename: string | null;
 	isFetched: boolean;
+	onChange: ( media: Partial< Attachment > ) => void;
 };
 
 export type MediaPreferences = {
